refactor(job): rusty and safe
