+++ src/config.rs
@@ -3,9 +3,9 @@ use strum_macros::{AsRefStr, Display, EnumIter, EnumProperty, EnumString};
 
 use crate::constants::DC_VERSION_STR;
 use crate::context::Context;
-use crate::dc_job::*;
 use crate::dc_tools::*;
 use crate::error::Error;
+use crate::job::*;
 use crate::stock::StockMessage;
 
 /// The available configuration keys.
@@ -102,17 +102,17 @@ impl Context {
             }
             Config::InboxWatch => {
                 let ret = self.sql.set_config(self, key, value);
-                unsafe { dc_interrupt_imap_idle(self) };
+                interrupt_imap_idle(self);
                 ret
             }
             Config::SentboxWatch => {
                 let ret = self.sql.set_config(self, key, value);
-                dc_interrupt_sentbox_idle(self);
+                interrupt_sentbox_idle(self);
                 ret
             }
             Config::MvboxWatch => {
                 let ret = self.sql.set_config(self, key, value);
-                dc_interrupt_mvbox_idle(self);
+                interrupt_mvbox_idle(self);
                 ret
             }
             Config::Selfstatus => {
