+++ src/context.rs
@@ -3,13 +3,13 @@ use std::sync::{Arc, Condvar, Mutex, RwLock};
 use crate::chat::*;
 use crate::constants::*;
 use crate::contact::*;
-use crate::dc_job::*;
 use crate::dc_loginparam::*;
 use crate::dc_move::*;
 use crate::dc_msg::*;
 use crate::dc_receive_imf::*;
 use crate::dc_tools::*;
 use crate::imap::*;
+use crate::job::*;
 use crate::job_thread::JobThread;
 use crate::key::*;
 use crate::lot::Lot;
@@ -230,7 +230,13 @@ unsafe fn cb_precheck_imf(
         }
         dc_do_heuristics_moves(context, server_folder, msg_id);
         if 0 != mark_seen {
-            dc_job_add(context, 130, msg_id as libc::c_int, Params::new(), 0);
+            job_add(
+                context,
+                Action::MarkseenMsgOnImap,
+                msg_id as libc::c_int,
+                Params::new(),
+                0,
+            );
         }
     }
     free(old_server_folder as *mut libc::c_void);
