+++ src/lib.rs
@@ -30,6 +30,7 @@ pub mod constants;
 pub mod contact;
 pub mod context;
 mod imap;
+pub mod job;
 mod job_thread;
 pub mod key;
 pub mod keyring;
@@ -50,7 +51,6 @@ pub mod dc_configure;
 mod dc_dehtml;
 mod dc_e2ee;
 pub mod dc_imex;
-pub mod dc_job;
 pub mod dc_location;
 mod dc_loginparam;
 mod dc_mimefactory;
